Refactor code into helper module
