+++ backend/requirements.txt
@@ -13,3 +13,4 @@ pytz==2021.1
 requests==2.22.0
 sqlparse==0.4.1
 urllib3==1.25.11
+cryptography==3.4.7
